requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
dj-database-url==1.0.0
Django==4.1
djangorestframework==3.13.1
gunicorn==20.1.0
PyMySQL==1.0.2
pytz==2022.2.1
sqlparse==0.4.2
whitenoise==6.2.0
